+++ packages/plugin-landing-page-graphql-playground/src/index.ts
@@ -5,11 +5,7 @@
 // specifying `version` when installing the plugin.
 
 import { renderPlaygroundPage } from '@apollographql/graphql-playground-html';
-import type {
-  ApolloServerPlugin,
-  BaseContext,
-  GraphQLServerListener,
-} from '../../../externalTypes';
+import type { ApolloServerPlugin, GraphQLServerListener } from '@apollo/server';
 
 // This specifies the React version of our fork of GraphQL Playground,
 // `@apollographql/graphql-playground-react`.  It is related to, but not to
@@ -27,13 +23,11 @@ export type ApolloServerPluginLandingPageGraphQLPlaygroundOptions = Parameters<
   typeof renderPlaygroundPage
 >[0];
 
-export function ApolloServerPluginLandingPageGraphQLPlayground<
-  TContext extends BaseContext,
->(
+export function ApolloServerPluginLandingPageGraphQLPlayground(
   options: ApolloServerPluginLandingPageGraphQLPlaygroundOptions = Object.create(
     null,
   ),
-): ApolloServerPlugin<TContext> {
+): ApolloServerPlugin {
   return {
     async serverWillStart(): Promise<GraphQLServerListener> {
       return {