+++ packages/server/src/runHttpQuery.ts
@@ -234,9 +234,9 @@ export function cloneObject<T extends Object>(object: T): T {
   return Object.assign(Object.create(Object.getPrototypeOf(object)), object);
 }
 
-export function newHTTPGraphQLHead(statusCode?: number): HTTPGraphQLHead {
+export function newHTTPGraphQLHead(status?: number): HTTPGraphQLHead {
   return {
-    statusCode,
+    status,
     headers: new HeaderMap(),
   };
 }