+++ scripts/postcompile.mjs
@@ -12,7 +12,11 @@ import { writeFileSync } from 'fs';
 import rimraf from 'rimraf';
 
 // Tell Node what kinds of files the ".js" files in these subdirectories are.
-for (const dir of ['plugin-response-cache', 'server']) {
+for (const dir of [
+  'plugin-response-cache',
+  'plugin-landing-page-graphql-playground',
+  'server',
+]) {
   writeFileSync(
     path.join('packages', dir, 'dist', 'esm', 'package.json'),
     JSON.stringify({ type: 'module' }),